More sounds

+++ src/Terrain.cpp
@@ -1,3 +1,4 @@
+#include <mutex>
 
 #include "BulletCollision/CollisionDispatch/btCollisionObjectWrapper.h"
 #include "BulletCollision/CollisionDispatch/btCollisionObject.h"
@@ -184,7 +185,7 @@ void Terrain::Chunk::prepareBuffers()
     constexpr glm::vec3 farPoint(10000.f, 0.f, 0.f);
 
     #pragma omp parallel for
-    for (size_t i = 0; i < texCoords.size() / 6; i++)
+    for (int i = 0; i < texCoords.size() / 6; i++)
     {
         glm::vec3 a(vertices[9 * i], vertices[9 * i + 1], vertices[9 * i + 2]);
         glm::vec3 b(vertices[9 * i + 3], vertices[9 * i + 4], vertices[9 * i + 5]);
